Cleanup config.
