+++ sample-isv-web/src/main/resources/application-local.yml
@@ -1,19 +1,20 @@
 server:
-    port: 9080
+  port: 9080
 
 spring:
-    datasource:
-        url: jdbc:mysql://localhost/isv
-        username: root
-        password: password
-        driver-class-name: com.mysql.jdbc.Driver
-    jpa:
-        generate-ddl: true
+  datasource:
+    url: jdbc:mysql://localhost/isv
+    username: root
+    password: password
+    driver-class-name: com.mysql.jdbc.Driver
+  jpa:
+    database-platform: org.hibernate.dialect.MySQL5InnoDBDialect
+    generate-ddl: true
 
 appdirect:
-    base:
-        url: http://local.appdirect.com
-    oauth:
-        consumer:
-            key: key
-            secret: secret
+  base:
+    url: http://local.appdirect.com
+  oauth:
+    consumer:
+      key: key
+      secret: secret
