+++ sample-isv-web/src/main/resources/application-testisv1.yml
@@ -1,16 +1,17 @@
 spring:
-    datasource:
-        url: jdbc:mysql://us-cdbr-east-05.cleardb.net:3306/ad_3fd6a735c149159
-        username: b4c9356d57d688
-        password: 95566d10
-        driver-class-name: com.mysql.jdbc.Driver
-    jpa:
-        generate-ddl: true
+  datasource:
+    url: jdbc:mysql://us-cdbr-east-05.cleardb.net:3306/ad_3fd6a735c149159
+    username: b4c9356d57d688
+    password: 95566d10
+    driver-class-name: com.mysql.jdbc.Driver
+  jpa:
+    database-platform: org.hibernate.dialect.MySQL5InnoDBDialect
+    generate-ddl: true
 
 appdirect:
-    base:
-        url: https://test.appdirect.com
-    oauth:
-        consumer:
-            key: demo-app-1
-            secret: fGWi8NZzjy0mbJLM
+  base:
+    url: https://test.appdirect.com
+  oauth:
+    consumer:
+      key: demo-app-1
+      secret: fGWi8NZzjy0mbJLM
